+++ tests/testthat/test-lp_nl_iv.R
@@ -22,7 +22,7 @@ context("check_input_lp_nl_iv")
 
     expect_error(lp_nl_iv(endog_data,
                               lags_endog_nl           = 3,
-                              instr             = instrument,
+                              shock             = instrument,
                               exog_data         = NULL,
                               lags_exog         = NULL,
                               contemp_data      = NULL,
@@ -43,7 +43,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether trend is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -65,7 +65,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether switching variable is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -86,7 +86,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether use_hp is set", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -107,7 +107,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whetherlambda is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -128,7 +128,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether gamma is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -149,7 +149,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether width for confidence bands is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -169,7 +169,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether number of horizons is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -189,7 +189,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether lag  length criterion is correctly spelled", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = NaN,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -210,7 +210,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether lags criterion and fixed number of lags for nonlinear model is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = 3,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -231,7 +231,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether maximum number of lags is given", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = NaN,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -251,7 +251,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether number of horizons is positive", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = NaN,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -272,7 +272,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether trend is correctly specified", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = NaN,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -294,7 +294,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether width of confidence bands is correctly specified", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = NaN,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -315,7 +315,7 @@ context("check_input_lp_nl_iv")
   test_that("Check whether gamma is negative number", {
     expect_error(lp_nl_iv(endog_data,
                           lags_endog_nl           = NaN,
-                          instr             = instrument,
+                          shock             = instrument,
                           exog_data         = NULL,
                           lags_exog         = NULL,
                           contemp_data      = NULL,
@@ -370,7 +370,7 @@ test_that("Compare results with RZ-2018", {
   # Estimate local projections
   results_nl_iv <- lp_nl_iv(endog_data,
                             lags_endog_nl     = 3,
-                            instr             = instrument,
+                            shock             = instrument,
                             exog_data         = exog_data,
                             lags_exog         = 4,
                             contemp_data      = NULL,