+++ tests/testthat/test-plot_lin.R
@@ -54,7 +54,7 @@ test_that("Check whether list length is consistent number of endogenous variable
           for lin_iv model", {
 
   results_lin_iv <- lp_lin_iv(endog_data,
-                               instr          = instrument,
+                               shock          = instrument,
                                lags_endog_lin = 4,
                                exog_data      = NULL,
                                lags_exog      = NULL,